# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * FamJam – multi-reunion family gathering platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface PlatformSettings {
  reunionCreationEnabled: boolean;
}

export type ShirtSize = typeof ShirtSize[keyof typeof ShirtSize];


export const ShirtSize = {
  XS: 'XS',
  S: 'S',
  M: 'M',
  L: 'L',
  XL: 'XL',
  '2XL': '2XL',
  '3XL': '3XL',
} as const;

export type PaymentStatus = typeof PaymentStatus[keyof typeof PaymentStatus];


export const PaymentStatus = {
  pending: 'pending',
  paid: 'paid',
  waived: 'waived',
} as const;

export interface ReunionBranch {
  id: number;
  reunionId: number;
  name: string;
  sortOrder: number;
}

export interface Reunion {
  id: number;
  code: string;
  name: string;
  /** ISO date (YYYY-MM-DD) */
  startDate: string;
  /** ISO date (YYYY-MM-DD) */
  endDate: string;
  feePerPerson: number;
  paymentHandle: string;
  /** @nullable */
  paymentUrl?: string | null;
  organizerId?: string;
  createdAt: string;
  branches: ReunionBranch[];
}

export interface ReunionSummary {
  reunion: Reunion;
  registrationCount: number;
  attendeeCount: number;
}

export interface ReunionInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  startDate: string;
  /** @minLength 1 */
  endDate: string;
  /** @minimum 0 */
  feePerPerson: number;
  /** @minLength 1 */
  paymentHandle: string;
  paymentUrl?: string;
  /**
     * @minItems 1
     * @items.minLength 1
     */
  branches: string[];
}

export interface ReunionUpdateInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  startDate: string;
  /** @minLength 1 */
  endDate: string;
  /** @minimum 0 */
  feePerPerson: number;
  /** @minLength 1 */
  paymentHandle: string;
  paymentUrl?: string;
}

export interface BranchInput {
  /** @minLength 1 */
  name: string;
  sortOrder?: number;
}

export interface AttendeeInput {
  /** @minLength 1 */
  name: string;
  shirtSize: ShirtSize;
  dietaryRestrictions?: string;
}

export interface Attendee {
  id: number;
  registrationId: number;
  name: string;
  shirtSize: ShirtSize;
  /** @nullable */
  dietaryRestrictions?: string | null;
}

export interface RegistrationInput {
  reunionId: number;
  /** @minLength 1 */
  branchName: string;
  /** @minItems 1 */
  attendees: AttendeeInput[];
}

export interface Registration {
  id: number;
  reunionId: number;
  /** @nullable */
  reunionName?: string | null;
  /** @nullable */
  reunionCode?: string | null;
  userId: string;
  branchName: string;
  attendeeCount: number;
  paymentStatus: PaymentStatus;
  createdAt: string;
  attendees: Attendee[];
}

export interface AdminRegistration {
  id: number;
  reunionId: number;
  userId: string;
  userEmail: string;
  /** @nullable */
  userName?: string | null;
  branchName: string;
  attendeeCount: number;
  paymentStatus: PaymentStatus;
  createdAt: string;
  attendees: Attendee[];
}

export interface UpdatePaymentStatusInput {
  paymentStatus: PaymentStatus;
}

export interface GroupCount {
  branchName: string;
  registrationCount: number;
  attendeeCount: number;
}

export interface RegistrationSummary {
  totalRegistrations: number;
  totalAttendees: number;
  byGroup: GroupCount[];
}

export interface ShirtSizeCount {
  shirtSize: ShirtSize;
  count: number;
}

export interface DayRegistrationCount {
  date: string;
  count: number;
}

export interface AdminReport {
  totalRegistrations: number;
  totalAttendees: number;
  paidCount: number;
  pendingCount: number;
  waivedCount: number;
  dietaryCount: number;
  byGroup: GroupCount[];
  byShirtSize: ShirtSizeCount[];
  registrationsOverTime: DayRegistrationCount[];
}

export interface Announcement {
  id: number;
  reunionId: number;
  title: string;
  body: string;
  createdAt: string;
  pinned?: boolean;
}

export interface AnnouncementInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  body: string;
  pinned?: boolean;
}

export interface ScheduleItem {
  id: number;
  reunionId: number;
  /** Day label, e.g. Thursday July 17 */
  day: string;
  /** Time string, e.g. 9:00 AM */
  startTime: string;
  /** @nullable */
  endTime?: string | null;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  location?: string | null;
  sortOrder?: number;
}

export interface ScheduleItemInput {
  /** @minLength 1 */
  day: string;
  /** @minLength 1 */
  startTime: string;
  endTime?: string;
  /** @minLength 1 */
  title: string;
  description?: string;
  location?: string;
  sortOrder?: number;
}

export interface AdminUser {
  id: string;
  email: string;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  isAdmin: boolean;
  registrationCount: number;
  attendeeCount: number;
  createdAt: string;
}

export interface ToggleAdminInput {
  isAdmin: boolean;
}