# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FamJam – multi-reunion family gathering platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /registrations:
    post:
      operationId: createRegistration
      tags: [registrations]
      summary: Submit a new reunion registration
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationInput"
      responses:
        "201":
          description: Registration created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
  /registrations/mine:
    get:
      operationId: listMyRegistrations
      tags: [registrations]
      summary: List the current user's registrations
      security:
        - clerkAuth: []
      responses:
        "200":
          description: List of registrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Registration"
        "401":
          description: Unauthorized
  /registrations/{id}:
    get:
      operationId: getRegistration
      tags: [registrations]
      summary: Get a single registration with attendees
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Registration detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "404":
          description: Not found

  # ── Platform admin ────────────────────────────────────────────────────────────
  /admin/settings:
    patch:
      operationId: adminUpdateSettings
      tags: [admin]
      summary: Enable or disable reunion creation platform-wide
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatformSettings"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettings"
        "403":
          description: Forbidden
  /admin/reunions:
    get:
      operationId: adminListReunions
      tags: [admin]
      summary: List all reunions across the platform
      security:
        - clerkAuth: []
      responses:
        "200":
          description: List of reunions with stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReunionSummary"
        "403":
          description: Forbidden
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all registered users with registration stats
      security:
        - clerkAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
        "403":
          description: Forbidden
  /admin/users/{id}/admin:
    patch:
      operationId: adminToggleAdminFlag
      tags: [admin]
      summary: Grant or revoke platform-admin access for a user
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleAdminInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "403":
          description: Forbidden
        "404":
          description: Not found
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    PlatformSettings:
      type: object
      required: [reunionCreationEnabled]
      properties:
        reunionCreationEnabled:
          type: boolean
    ShirtSize:
      type: string
      enum: [XS, S, M, L, XL, "2XL", "3XL"]
    PaymentStatus:
      type: string
      enum: [pending, paid, waived]
    ReunionBranch:
      type: object
      required: [id, reunionId, name, sortOrder]
      properties:
        id:
          type: integer
        reunionId:
          type: integer
        name:
          type: string
        sortOrder:
          type: integer
    Reunion:
      type: object
      required:
        [id, code, name, startDate, endDate, feePerPerson, paymentHandle, createdAt, branches]
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        startDate:
          type: string
          description: "ISO date (YYYY-MM-DD)"
        endDate:
          type: string
          description: "ISO date (YYYY-MM-DD)"
        feePerPerson:
          type: integer
        paymentHandle:
          type: string
        paymentUrl:
          type: ["string", "null"]
        organizerId:
          type: string
        createdAt:
          type: string
          format: date-time
        branches:
          type: array
          items:
            $ref: "#/components/schemas/ReunionBranch"
    ReunionSummary:
      type: object
      required: [reunion, registrationCount, attendeeCount]
      properties:
        reunion:
          $ref: "#/components/schemas/Reunion"
        registrationCount:
          type: integer
        attendeeCount:
          type: integer
...
    AttendeeInput:
      type: object
      required: [name, shirtSize]
      properties:
        name:
          type: string
          minLength: 1
        shirtSize:
          $ref: "#/components/schemas/ShirtSize"
        dietaryRestrictions:
          type: string
    Attendee:
      type: object
      required: [id, registrationId, name, shirtSize]
      properties:
        id:
          type: integer
        registrationId:
          type: integer
        name:
          type: string
        shirtSize:
          $ref: "#/components/schemas/ShirtSize"
        dietaryRestrictions:
          type: ["string", "null"]
    RegistrationInput:
      type: object
      required: [reunionId, branchName, attendees]
      properties:
        reunionId:
          type: integer
        branchName:
          type: string
          minLength: 1
        attendees:
          type: array
          items:
            $ref: "#/components/schemas/AttendeeInput"
          minItems: 1
    Registration:
      type: object
      required:
        [id, reunionId, userId, branchName, attendeeCount, createdAt, attendees, paymentStatus]
      properties:
        id:
          type: integer
        reunionId:
          type: integer
        reunionName:
          type: ["string", "null"]
        reunionCode:
          type: ["string", "null"]
        userId:
          type: string
        branchName:
          type: string
        attendeeCount:
          type: integer
        paymentStatus:
          $ref: "#/components/schemas/PaymentStatus"
        createdAt:
          type: string
          format: date-time
        attendees:
          type: array
          items:
            $ref: "#/components/schemas/Attendee"
...
    AdminUser:
      type: object
      required: [id, email, isAdmin, createdAt, registrationCount, attendeeCount]
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        isAdmin:
          type: boolean
        registrationCount:
          type: integer
        attendeeCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    ToggleAdminInput:
      type: object
      required: [isAdmin]
      properties:
        isAdmin:
          type: boolean