# lib/api-zod/src/generated/types/attendeeInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * FamJam – multi-reunion family gathering platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ShirtSize } from './shirtSize';

export interface AttendeeInput {
  /** @minLength 1 */
  name: string;
  shirtSize: ShirtSize;
  dietaryRestrictions?: string;
}