# lib/api-zod/src/generated/types/registrationInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * FamJam – multi-reunion family gathering platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AttendeeInput } from './attendeeInput';

export interface RegistrationInput {
  reunionId: number;
  /** @minLength 1 */
  branchName: string;
  /** @minItems 1 */
  attendees: AttendeeInput[];
}